# middleware/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the Open-Xchange Admin Panel API. This API is the connector\
    \ to the middleware that handles provisioning tasks towards Open-Xchange products\
    \ such as Dovecot or App Suite."
  version: "1.0.0"
  title: "Open-Xchange Admin Panel API"
  termsOfService: "http://www.open-xchange.com"
  contact:
    name: "Benjamin Otterbach"
    email: "bartl3by@t-online.de"
  license:
    name: "GPL3"
    url: "https://github.com/bartl3by/oxap/blob/master/LICENSE"
basePath: "/v2"
...
schemes:
- "http"
- "https"
paths:
...
  /oxap/session:
    get:
      tags:
      - "oxapaccount"
      summary: "Refresh and get OXAP session information"
      description: "Refresh and get OXAP session information"
      operationId: "get_session_information"
      produces:
      - "application/json"
      parameters:
      - name: "Oxapsessionid"
        in: "header"
        description: "The session id previously acquired from /oxap/session."
        required: false
        type: "string"
      responses:
        200:
          description: "Successful Operation"
          schema:
            $ref: "#/definitions/Session"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-security-scope: ""
      x-swagger-router-controller: "swagger_server.controllers.oxapaccount_controller"
    post:
      tags:
      - "oxapaccount"
      summary: "Create an OXAP session"
      description: "Create an OXAP session. The endpoint_id parameter is mandatory\
        \ for 'context', 'reseller' and 'user' login requests (set via 'role'), a\
        \ login request will fail if this parameter is not specified in case the role\
        \ is 'context', 'reseller' or 'user'. OXAP accounts (role 'oxap') do not need\
        \ to specify this value as a session grants them access to all their endpoints,\
        \ endpoints will have to be specified during subsequent API requests to OXAP\
        \ interfaces (like the manage context interface).\n"
      operationId: "create_session"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "data"
        required: true
        schema:
          $ref: "#/definitions/SessionLogin"
      responses:
        200:
          description: "Successful Operation"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        422:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.oxapaccount_controller"
    delete:
      tags:
      - "oxapaccount"
      summary: "Delete an OXAP session"
      description: "Delete an OXAP session"
      operationId: "delete_session"
      produces:
      - "application/json"
      parameters:
      - name: "Oxapsessionid"
        in: "header"
        description: "The session id previously acquired from /oxap/session."
        required: false
        type: "string"
      responses:
        200:
          description: "Successful Operation"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-security-scope: ""
      x-swagger-router-controller: "swagger_server.controllers.oxapaccount_controller"
definitions:
  Error:
    type: "string"
    description: "The cause of the Error"
  SessionLogin:
    type: "object"
    required:
    - "account_id"
    - "password"
    - "username"
    properties:
      username:
        type: "string"
        description: "Id of the OXAP account"
      password:
        type: "string"
        format: "password"
        description: "Id of the OXAP endpoint for this OXAP account"
      account_id:
        type: "string"
        description: "Id of the OXAP account"
      endpoint_id:
        type: "string"
        description: "Id of the OXAP endpoint for this OXAP account"
      role:
        type: "string"
        description: "'The type of account that requests the login. If type is ':\
          \ * oxap - An OXAP Master administrator having access to multiple endpoints\
          \ * reseller - A Context reseller managing a context within an endpoint\
          \ (the endpoint_id is required in data object) * context - A Context administrator\
          \ managing a context within an endpoint (the endpoint_id is required in\
          \ data object) * user - An Open-Xchange user account (the endpoint_id is\
          \ required in data object)"
        enum:
        - "oxap"
        - "context"
        - "reseller"
        - "user"
    description: "Session Login Information"
    example:
      password: "password"
      account_id: "account_id"
      role: "oxap"
      endpoint_id: "endpoint_id"
      username: "username"
  Session:
    type: "object"
    required:
    - "id"
    properties:
      id:
        type: "string"
        description: "The OXAP Session ID generated by the OXAP middleware"
      creation_date:
        type: "string"
      expiration_date:
        type: "string"
      account_id:
        type: "string"
        description: "Id of the OXAP account"
      endpoint_id:
        type: "string"
        description: "Id of the OXAP endpoint for this OXAP account"
      role:
        type: "string"
        description: "'The type of account that requests the login. If type is ':\
          \ * oxap - An OXAP Master administrator having access to multiple endpoints\
          \ * reseller - A Context reseller managing a context within an endpoint\
          \ (the endpoint_id is required in data object) * context - A Context administrator\
          \ managing a context within an endpoint (the endpoint_id is required in\
          \ data object) * user - An Open-Xchange user account (the endpoint_id is\
          \ required in data object)"
        enum:
        - "oxap"
        - "context"
        - "reseller"
        - "user"
    description: "OXAP Session"
    example:
      account_id: "account_id"
      role: "oxap"
      endpoint_id: "endpoint_id"
      id: "id"
      creation_date: "creation_date"
      expiration_date: "expiration_date"
...